cleared out a few things
